refactor(core): move computed label/grabbedBox accessors into labelManager
The two viewport-aware projections (computedLabelInstances,
computedGrabbedBoxes) plus the supporting `recomputeLabelInstance`
helper and labelInstanceCache map were sitting at ~line 2330 of
init(), far from the label state they project over.

Move all three into createLabelInstanceManager and expose
`computedLabelInstances` + `computedGrabbedBoxes` from its
returned API. The manager now owns the entire label/grabbed-box
state lifecycle: creation, fade timers, hover behavior, and
renderer projection.

The manager input is now a named-args object that adds the two
theme-enabled accessors needed by the computed memos. The previous
`createLabelInstanceManager(grab, pluginRegistry)` positional
signature is replaced with `createLabelInstanceManager({ grab,
pluginRegistry, isThemeEnabled, isGrabbedBoxesThemeEnabled })`.

index.tsx: 3018 -> 2921 (-97). All 57 copy-feedback +
context-menu specs green.
